docs: describe Nicholson default workflow
