+++ README.md
@@ -1,6 +1,6 @@
 # VideoCut
 
-VideoCut provides a small video editing pipeline driven by WhisperX transcripts.  Each step of the process is broken out into a script under `scripts/` and backed by small modules for re-use.
+VideoCut provides a small video editing pipeline driven by WhisperX transcripts.  The project now exposes a single `videocut` command powered by **Typer** for easy usage.  All functionality can still be invoked step-by-step or via a one-shot pipeline command.
 
 ## Requirements
 
@@ -26,41 +26,36 @@ HF_TOKEN=your_hf_token_here
 
 ## Workflow
 
-1. **Transcribe** – `scripts/transcribe_step.py` runs WhisperX and produces `markup_guide.txt` and the original JSON file.
-2. **Prepare for editing** – use `scripts/json_to_editable_step.py` to create `segments_edit.json`.
-3. **Identify clips** – after marking segments to keep, run
-   `scripts/identify_clips_json_step.py` or
-   `scripts/extract_marked_step.py` to create `segments_to_keep.json`.
-4. **Auto-select Nicholson segments (optional)** – `scripts/auto_mark_nicholson_step.py` examines a diarized JSON file and builds the JSON list for you.
-5. **Generate clips** – `scripts/generate_clips_step.py` cuts clips from the input video into a `clips/` directory.
-6. **Concatenate** – `scripts/concatenate_clips_step.py` stitches the clips together with white flashes between each.
+1. **Transcribe with diarization** – `videocut transcribe input.mp4 --diarize --hf_token $HF_TOKEN` runs WhisperX and produces `markup_guide.txt` and `input.json` with speaker labels.
+2. **Auto-mark Nicholson** – `videocut auto-mark-nicholson input.json` generates `segments_to_keep.json` selecting Secretary Nicholson's speech.
+3. **Review and edit** – optionally run `videocut json-to-editable segments_to_keep.json` and modify the JSON to fine‑tune the clips.
+4. **Generate clips** – `videocut generate-clips input.mp4` cuts clips into a `clips/` directory.
+5. **Concatenate** – `videocut concatenate` stitches the clips together with white flashes.
 
-All of these steps can be executed sequentially with `scripts/run_pipeline.py --all`.
+All of these steps can be executed sequentially with `videocut pipeline input.mp4 --diarize --hf_token $HF_TOKEN` which auto‑marks Nicholson by default.
 
 ### Example commands
 
 ```bash
 # Transcription with diarization
-python3 scripts/transcribe_step.py --input meeting.mp4 --diarize --hf_token $HF_TOKEN
+videocut transcribe meeting.mp4 --diarize --hf_token $HF_TOKEN
 
-# Create an editable JSON
-python3 scripts/json_to_editable_step.py meeting.json --out segments_edit.json
-# (edit segments_edit.json to mark segments to keep)
+# Auto-mark Nicholson segments
+videocut auto-mark-nicholson meeting.json
 
-# Generate the list of segments
-python3 scripts/identify_clips_json_step.py --json segments_edit.json
+# (Optional) tweak the segments
+videocut json-to-editable segments_to_keep.json --out segments_edit.json
+# ...edit segments_edit.json as desired...
 
 # Cut clips and assemble the final video
-python3 scripts/generate_clips_step.py --input meeting.mp4 --json segments_to_keep.json
-python3 scripts/concatenate_clips_step.py --clips_dir clips --out final.mp4
+videocut generate-clips meeting.mp4 --segs segments_to_keep.json
+videocut concatenate --clips_dir clips --out final.mp4
 ```
 
 ## Package layout
 
-- `transcribe.py` – wrapper around WhisperX used by multiple scripts
-- `clip_utils.py` – helpers for converting transcripts, identifying clips, cutting and concatenating video
-- `scripts/` – single-purpose step scripts (see above) including `run_pipeline.py`
-- `process_video.py` and `lightweight_process.py` – monolithic CLIs combining all steps
+- `videocut/cli.py` – Typer command line interface
+- `videocut/core/` – modular helpers (`transcription.py`, `segmentation.py`, `video_editing.py`, `nicholson.py`)
 - `videos/` – example data used for testing the pipeline
 
-WhisperX and FFmpeg must be installed separately.  Once those are available, the scripts above can be combined or run individually to automate cutting long meeting videos into polished clips.
+WhisperX and FFmpeg must be installed separately.  Once those are available, the `videocut` command can automate cutting long meeting videos into polished clips.
