+++ requirements.txt
@@ -1,3 +1,4 @@
 # WhisperX pulls in its own dependencies
 python-dotenv==1.1.0
 whisperx==3.3.4
+typer==0.12.3
